# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nexus Pro Backend API
  version: 0.2.0
  description: |
    Identifiers for persisted tenant resources are globally unique across tenants.
    tenant_id scopes ownership, authorization, and row-level security.
security:
  - bearerAuth: []
paths:
...
  /v1/workspace:
    get:
      summary: Get workspace settings aggregate
      responses:
        "200":
          description: Workspace settings panes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformWorkspaceDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /v1/workspace/agent-models:
    get:
      summary: List workspace agent models
      responses:
        "200":
          description: Agent model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalError"}
    post:
      summary: Create workspace agent model
      requestBody: {required: true, content: {application/json: {schema: {type: object}}}}
      responses:
        "201":
          description: Created agent model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agent-models/{id}:
    patch:
      summary: Update workspace agent model
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody: {required: true, content: {application/json: {schema: {type: object}}}}
      responses:
        "200":
          description: Updated agent model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
    delete:
      summary: Delete workspace agent model
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Deleted agent model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agent-models/{id}/sync:
    post:
      summary: Sync workspace agent model to LiteLLM
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Synced agent model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "502": {$ref: "#/components/responses/InternalError"}
        "503": {$ref: "#/components/responses/InternalError"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agent-models/{id}/test:
    post:
      summary: Test workspace agent model connection
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Tested agent model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agents:
    get:
      summary: List workspace agents
      responses:
        "200":
          description: Agent definition list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalError"}
    post:
      summary: Create workspace agent
      requestBody: {required: true, content: {application/json: {schema: {type: object}}}}
      responses:
        "201":
          description: Created agent definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agents/{id}:
    patch:
      summary: Update workspace agent
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody: {required: true, content: {application/json: {schema: {type: object}}}}
      responses:
        "200":
          description: Updated agent definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
    delete:
      summary: Delete workspace agent
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Deleted agent definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agents/{id}/publish:
    post:
      summary: Publish workspace agent
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Published agent definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agents/{id}/unpublish:
    post:
      summary: Stop publishing workspace agent
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Unpublished agent definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agents/{id}/trial:
    post:
      summary: Trial workspace agent
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody: {required: true, content: {application/json: {schema: {type: object}}}}
      responses:
        "200":
          description: Agent trial result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agents/{id}/rollback:
    post:
      summary: Roll back workspace agent to a previous version
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody: {required: true, content: {application/json: {schema: {type: object}}}}
      responses:
        "200":
          description: Rolled-back agent definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workspace/agents/tools:
    get:
      summary: List workspace agent tools
      responses:
        "200":
          description: Agent tools
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalError"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token issued by the configured identity provider and mapped to a local tenant account.
  parameters:
    ID:
      name: id
      in: path
      required: true
      description: Globally unique persisted resource identifier; tenant_id still scopes ownership and authorization.
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Validation failed. Field errors include tab, field, code and message; import row errors include row_number.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: Authenticated tenant/account context is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Permission, menu, button, field, or data-scope access was denied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource was not found in the current tenant.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error. Response includes trace_id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    DataResponse:
      type: object
      required: [data]
      properties:
        data:
          description: Successful response payload.
...
    PlatformWorkspaceDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/PlatformWorkspaceResponse"
...
    PlatformWorkspaceResponse:
      type: object
      additionalProperties: true
...
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        tab:
          type: string
          description: Employee modal tab or section key, for example basic_info or employment_info.
        field:
          type: string
          description: Request or spreadsheet field key that failed validation.
        code:
          type: string
          description: Stable machine-readable validation or field-policy code.
        message:
          type: string
    RowError:
      type: object
      required: [row_number, field_errors]
      properties:
        row_number:
          type: integer
          minimum: 1
          description: Spreadsheet or CSV row number, including the header row offset used by the import parser.
        field_errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              type: integer
              format: int32
              description: |
                Numeric application error code. Current ranges:
                1xxxx common platform, request and authentication errors;
                2xxxx IAM and authorization errors; 3xxxx people-domain and HR errors;
                4xxxx attendance errors; 5xxxx workflow errors; 6xxxx agent errors.
              enum:
                - 10000
                - 10001
                - 10002
                - 10003
                - 10004
                - 10005
                - 10006
                - 10007
                - 10008
                - 10009
                - 10030
                - 10031
                - 10032
                - 10033
                - 10034
                - 10035
                - 10036
                - 10037
                - 10050
                - 10060
                - 10070
                - 20040
                - 20041
                - 20042
                - 20043
                - 20044
                - 20045
                - 30010
                - 30011
                - 30012
                - 30013
                - 30014
                - 30015
                - 30016
                - 30017
                - 30018
                - 30030
                - 30031
                - 30040
                - 30041
                - 30042
            message: {type: string}
            reason_code:
              type: string
              enum:
                - account_inactive
                - sso_email_not_authorized
                - sso_email_unverified
                - company_inactive
                - google_login_failed
                - sso_identity_conflict
                - sso_email_ambiguous
                - permission_missing
                - menu_denied
                - button_denied
                - field_denied
                - data_scope_denied
            field_errors:
              type: array
              items:
                $ref: "#/components/schemas/FieldError"
            row_errors:
              type: array
              items:
                $ref: "#/components/schemas/RowError"
            trace_id: {type: string}